Affichage lignes refusées

+++ gsbMVC/vues/v_detailsFicheVisiteur.php
@@ -41,17 +41,27 @@
        </thead>
 
        <tbody>
-          <?php
-              $totalFraisHorsForfait = 0;
-              foreach($lesFraisHorsForfait as $unFraisHorsForfait)
-              {
-                  $totalFraisHorsForfait += $unFraisHorsForfait['montant'];
-                  echo '<tr><td>'.$unFraisHorsForfait['date'].'</td>';
-                  echo '<td>'.$unFraisHorsForfait['libelle'].'</td>';
-                  echo '<td>'.$unFraisHorsForfait['montant'].'</td>';
-                  echo '<td><a href="index.php?uc=validationFrais&amp;action=deleteFicheHorsForfait&amp;ficheIDToDelete='.$unFraisHorsForfait['id'].'">Supprimer la ligne</a></td>';
-              }
-          ?>
+          <tr>
+             <?php
+                 $totalFraisHorsForfait = 0;
+                 foreach($lesFraisHorsForfait as $unFraisHorsForfait)
+                 {
+                     echo '<tr><td>'.$unFraisHorsForfait['date'].'</td>';
+                     echo '<td>'.$unFraisHorsForfait['libelle'].'</td>';
+                     echo '<td>'.$unFraisHorsForfait['montant'].'</td>';
+
+                     if (!strstr($unFraisHorsForfait['libelle'], 'REFUSE'))
+                     {
+                        $totalFraisHorsForfait += $unFraisHorsForfait['montant'];
+                        echo '<td><a href="index.php?uc=validationFrais&amp;action=deleteFicheHorsForfait&amp;ficheIDToDelete='.$unFraisHorsForfait['id'].'">Supprimer la ligne</a></td>';
+                     }
+                     else
+                     {
+                        echo '<td>/</td>';
+                     }
+                 }
+             ?>
+         </tr>
        </tbody>
     </table>
